+++ docs/tr2/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased](https://github.com/LostArtefacts/TRX/compare/tr2-0.10...develop) - ××××-××-××
+- added the bonus level game flow type, which allows for levels to be unlocked if all main game secrets are found (#2668)
 - fixed the final two levels not allowing for secrets to be counted in the statistics (#1582)
 - removed the need to specify in the game flow levels that have no secrets (secrets will be automatically counted) (#1582)